+++ shell_scripts/authorize_cloudsmith.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+echo $1 | dart pub token add https://dart.cloudsmith.io/alkami/fluttercore/
+echo $1 | dart pub token add https://dart.cloudsmith.io/alkami/flutterpackages/