+++ packages/webview_flutter/webview_flutter/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 3.1.0
+
+- chore:DEV-139004 publish to cloudsmith
+
+# 4.0.0
+
+- chore:DEV-139004 publish to cloudsmith
+
 ## 3.0.1
 
 * Removes a duplicate Android-specific integration test.