+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 3.1.0
+
+- chore:DEV-139004 publish to cloudsmith
+
+# 3.0.0
+
+- chore:DEV-139004 publish to cloudsmith
+
 ## NEXT
 
 * Fixes an integration test race condition.